⛑️ test : 인증 관련 픽스처 인증 픽스처 클래스로 이동
