+++ src/test/java/nextstep/member/acceptance/AuthAcceptanceSteps.java
@@ -1,11 +1,11 @@
 package nextstep.member.acceptance;
 
-import static nextstep.member.acceptance.MemberSteps.베어러_인증_로그인_요청;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
+import java.util.HashMap;
 import java.util.Map;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -32,6 +32,29 @@ public class AuthAcceptanceSteps {
                 .body(params)
                 .when().post("/login/github")
                 .then().log().all()
+                .extract();
+    }
+
+    public static ExtractableResponse<Response> 베어러_인증_로그인_요청(String email, String password) {
+        Map<String, String> params = new HashMap<>();
+        params.put("email", email);
+        params.put("password", password);
+
+        return RestAssured.given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(params)
+                .when().post("/login/token")
+                .then().log().all()
+                .extract();
+    }
+
+    public static ExtractableResponse<Response> 베이직_인증으로_내_회원_정보_조회_요청(String username, String password) {
+        return RestAssured.given().log().all()
+                .auth().preemptive().basic(username, password)
+                .accept(MediaType.APPLICATION_JSON_VALUE)
+                .when().get("/members/me")
+                .then().log().all()
+                .statusCode(HttpStatus.OK.value())
                 .extract();
     }
 }
